requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.0.12
dj-database-url==0.5.0
Django==4.0.5
django-heroku==0.3.1
django-taggit==3.0.0
Faker==13.15.0
gunicorn==20.1.0
idna==3.3
Pillow==9.2.0
psycopg2==2.9.3
python-dateutil==2.8.2
requests==2.28.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
whitenoise==6.2.0